// src/app/informe-auditoria/style-template/models/style-template.DTO.ts
export interface StyleTemplateDTO {
    id: number;
    nombre: string;
    fontFamily: string;
    fontSize: string;
    fontWeight: string;
    fontStyle: string;
    color: string;
    backgroundColor: string;
    lineHeight: string;
    textAlign: string;
    textDecoration: string;
    marginTop: string;
    marginBottom: string;
    marginLeft: string;
    marginRight: string;
    paddingTop: string;
    paddingBottom: string;
    paddingLeft: string;
    paddingRight: string;
    border: string;
    borderRadius: string;
  }